fix: 94 - [tour] code optimization and tour body style enhancements
