+++ blocks/badge/badge.js
@@ -1,4 +1,4 @@
-import { getParents } from '../../scripts/scripts.js';
+import { wrapMergeBlocksSection } from '../../scripts/scripts.js';
 
 function styleBlock(block) {
   const badgeElements = block.querySelector('a').children;
@@ -7,17 +7,6 @@ function styleBlock(block) {
     badgeElements[1].classList.add('upper-box');
     badgeElements[2].classList.add('lower-box');
   }
-}
-
-function rearrangeMergeSection(mergeBlockSection) {
-  const fragment = document.createElement('div');
-  mergeBlockSection.classList.remove('merge-blocks-desktop');
-  fragment.classList.add('merge-blocks-desktop');
-  const blocks = [...mergeBlockSection.children];
-  blocks.forEach((block) => {
-    fragment.appendChild(block);
-  });
-  mergeBlockSection.appendChild(fragment);
 }
 
 export default function decorate(block) {
@@ -36,8 +25,8 @@ export default function decorate(block) {
   }
   styleBlock(block);
   // extract merge-blocks-desktop elements in a dedicated div
-  const mergeBlocks = getParents(block, '.merge-blocks-desktop');
-  if (mergeBlocks.length > 0) {
-    rearrangeMergeSection(mergeBlocks[0]);
+  const mergeBlocks = block.closest('.merge-blocks-desktop');
+  if (mergeBlocks != null) {
+    wrapMergeBlocksSection(mergeBlocks);
   }
 }
