+++ blocks/header/header.js
@@ -4,7 +4,6 @@ import { fetchNavigationConfig } from '../../scripts/scripts.js';
 
 export default async function decorate(block) {
   const data = await fetchNavigationConfig();
-  console.dir('Navigation config', data);
 
   let lastScroll = 0;
   document.addEventListener('scroll', () => {
