+++ styles/styles.css
@@ -320,13 +320,23 @@ body.tour main p {
 }
 
 body.tour main .section {
-  padding: 0;
+  padding: 0 20px;
   position:relative;
   background: var(--background-color);
 }
 
 body.tour main .section p {
   margin-bottom: 0;
+}
+
+/* stylelint-disable-next-line no-descending-specificity */
+body.tour .section .default-content-wrapper {
+  padding-bottom: 61px;
+  padding-top: 54px;
+}
+
+body.tour .section.padding-bottom-small .default-content-wrapper {
+  padding-bottom: 28px;
 }
 
 body.tour .default-content-wrapper ul {
@@ -342,7 +352,6 @@ body.tour main .section-organize-visit {
 body.tour main .section .default-content-wrapper p {
   color: #444;
   margin: 0;
-  padding-bottom: 28px;
 }
 
 /* stylelint-disable-next-line no-descending-specificity */
@@ -365,7 +374,12 @@ body.tour main .merge-blocks-desktop {
 }
 
 @media (min-width: 990px) {
-  body.tour .merge-blocks-desktop {
+  body.tour .section .default-content-wrapper {
+    padding-bottom: 89px;
+    padding-top: 83px;
+  }
+
+  body.tour main .merge-blocks-desktop {
     display: flex;
     width: var(--wrapper-width);
     margin: 0 auto;
